requirements.txt
ultralytics
torchvision
torch
mlflow
prefect
dvc
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
pandas
bentoml
prometheus-fastapi-instrumentator
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
confluent-kafka
pytest
ruff
onnx
onnxruntime
onnxslim
boto3
